# docs/swagger.yaml
definitions:
  domain.OrderStatus:
    enum:
    - UNDEFINDED
    - OPEN
    - CANCELLED
    - PENDING
    - RECEIVED
    - PREPARING
    - READY
    - COMPLETED
    type: string
    x-enum-varnames:
    - UNDEFINDED
    - OPEN
    - CANCELLED
    - PENDING
    - RECEIVED
    - PREPARING
    - READY
    - COMPLETED
...
  handler.UpdateOrder:
    properties:
      staff_id:
        example: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/domain.OrderStatus'
        example: OPEN, CANCELLED, PENDING, RECEIVED, PREPARING, READY, COMPLETED
    type: object
  handler.UpdateProduct:
    properties:
      category_id:
        example: 1
        type: integer
      description:
        example: The best burger in the world
        type: string
      name:
        example: McDonald's Big Mac
        type: string
      price:
        example: 29.9
        type: number
    type: object
...
  handler.createProductRequest:
    properties:
      category_id:
        example: 1
        type: integer
      description:
        example: The best burger in the world
        type: string
      name:
        example: BK Mega Stacker 2.0
        type: string
      price:
        example: 29.9
        type: number
    required:
    - category_id
    - description
    - name
    - price
    type: object
...
  request.CreateOrderRequest:
    properties:
      customer_id:
        example: 1
        type: integer
    required:
    - customer_id
    type: object
  request.UpdateOrderProductRequest:
    properties:
      id:
        example: 1
        type: integer
      quantity:
        example: 10
        type: integer
    required:
    - id
    - quantity
    type: object
...
  response.CategoryResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      errors:
        example:
        - '[''Validation failed: field X is required'''
        - ' ''Invalid format for field Y'']'
        items:
          type: string
        type: array
    type: object
...
  response.OrderPaginated:
    properties:
      limit:
        type: integer
      orders:
        items:
          $ref: '#/definitions/response.OrderResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  response.OrderProductPaginated:
    properties:
      limit:
        type: integer
      order_products:
        items:
          $ref: '#/definitions/response.OrderProductResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  response.OrderProductResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  response.OrderResponse:
    properties:
      created_at:
        type: string
      customerId:
        type: integer
      id:
        type: integer
      status:
        type: string
      totalBill:
        type: number
      updated_at:
        type: string
    type: object
...
  response.ProductPaginated:
    properties:
      limit:
        type: integer
      page:
        type: integer
      products:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      total:
        type: integer
    type: object
  response.ProductResponse:
    properties:
      active:
        type: boolean
      category:
        $ref: '#/definitions/response.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API para o Tech Challenge da FIAP - Fase 1 - G20 - 10 SOAT
  title: FIAP Tech Challenge Fase 1 - G20 - 10 SOAT
  version: "1"
paths:
...
  /api/v1/orders:
    get:
      consumes:
      - application/json
      description: List orders
      parameters:
      - description: OrderResponse name
        in: query
        name: name
        type: string
      - description: Customer ID
        in: query
        name: customer_id
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderPaginated'
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order
      parameters:
      - description: OrderResponse
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create an order
      tags:
      - orders
  /api/v1/orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order
      parameters:
      - description: OrderResponse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get an order
      parameters:
      - description: OrderResponse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get an order
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an order
      parameters:
      - description: OrderResponse ID
        in: path
        name: id
        required: true
        type: integer
      - description: OrderResponse
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an order
      tags:
      - orders
...
  /api/v1/orders/products:
    get:
      consumes:
      - application/json
      description: List order products
      parameters:
      - description: Order ID
        in: query
        name: order_id
        type: integer
      - description: Product ID
        in: query
        name: product_id
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderProductPaginated'
      summary: List order products
      tags:
      - orderHistories
    post:
      consumes:
      - application/json
      description: Update an order product
      parameters:
      - description: OrderProductResponse
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrderProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.OrderProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an order product
      tags:
      - orders
  /api/v1/orders/products/{id}:
    get:
      consumes:
      - application/json
      description: Get an order history
      parameters:
      - description: OrderHistoryResponse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get an order history
      tags:
      - orderHistories
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: List products
      parameters:
      - description: ProductResponse name
        in: query
        name: name
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductPaginated'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a product
      parameters:
      - description: ProductResponse
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.createProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product
      parameters:
      - description: ProductResponse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get a product
      parameters:
      - description: ProductResponse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: ProductResponse ID
        in: path
        name: id
        required: true
        type: integer
      - description: ProductResponse
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a product
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"